+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Some examples, more below in the actual changelog (newer entries are more likely
 * (internal) generic client: add hook FilterRequestURLHook (#123, @marioreggiori)
 
 -->
+
+* go-anxcloud is now tested with Go versions 1.20 and 1.21 and we might start using features only available since 1.20
+
 ## [0.5.3] - 2023-06-16
 
 ### Added